+++ examples/gcoap_dtls/client.c
@@ -1,0 +1,1 @@
+../gcoap/client.c
