--- examples/gcoap_dtls/gcoap_cli.c
@@ -1,1 +1,0 @@
-../gcoap/gcoap_cli.c
