+++ examples/gcoap_dtls/gcoap_example.h
@@ -1,0 +1,1 @@
+../gcoap/gcoap_example.h
