+++ examples/gcoap_dtls/server.c
@@ -1,0 +1,1 @@
+../gcoap/server.c
